+++ src/main/scala/net/akouryy/kreins/game/Pos.scala
@@ -1,18 +1,18 @@
 package net.akouryy.kreins.game
 
 object Pos {
-  def rotate180(p: Int) = {
-    ~p & 63
+  def rotate180(p: Byte) = {
+    (~p & 63).toByte
     // 0,0->7,7; 0,1->7,6; 4,2->3,5
   }
 
-  def mirrorWithDiagRightUp(p: Int) = {
-    (~p & 7) << 3 | ~p >> 3 & 7
+  def mirrorWithDiagRightUp(p: Byte) = {
+    ((~p & 7) << 3 | ~p >> 3 & 7).toByte
     // 0,0->7,7; 0,1->6,7; 4,2->5,3
   }
 
-  def mirrorWithDiagRightDown(p: Int) = {
-    (p & 7) << 3 | p >> 3
+  def mirrorWithDiagRightDown(p: Byte) = {
+    ((p & 7) << 3 | p >> 3).toByte
     // 0,0->0,0; 0,1->1,0; 4,2->2,4
   }
 }
